Design

+++ app/src/main/java/com/example/kelly/logeasyfinal/ScoreboardClass.java
@@ -6,8 +6,8 @@ import android.os.Parcelable;
 /**
  * Created by oanacozma on 04/04/15.
  */
-public class ScoreboardClass implements Parcelable {
-    private int user_id;
+public class ScoreboardClass implements Parcelable{
+    private long user_id;
     private int points;
     private int wrong_percent;
     private String level_id;
@@ -23,14 +23,14 @@ public class ScoreboardClass implements Parcelable {
         readFromParcel(in);
     }
 
-    public ScoreboardClass(int user, int no_of_points, int percent_wrong, String level){
+    public ScoreboardClass(long user, int no_of_points, int percent_wrong, String level){
         user_id=user;
         points=no_of_points;
         wrong_percent=percent_wrong;
         level_id=level;
     }
 
-    public int getUser_id() {
+    public long getUser_id() {
         return user_id;
     }
 
@@ -46,7 +46,7 @@ public class ScoreboardClass implements Parcelable {
         return level_id;
     }
 
-    public void setUser_id(int user_id) {
+    public void setUser_id(long user_id) {
         this.user_id = user_id;
     }
 
@@ -77,7 +77,7 @@ public class ScoreboardClass implements Parcelable {
     }
 
     private void readFromParcel(Parcel in) {
-        user_id = in.readInt();
+        user_id = in.readLong();
         points = in.readInt();
         wrong_percent = in.readInt();
         level_id = in.readString();
